Moved implicits queries to lowering/implicits module
