+++ crates/cairo-lang-lowering/src/implicits/mod.rs
@@ -5,9 +5,13 @@ use cairo_lang_defs::ids::LanguageElementId;
 use cairo_lang_diagnostics::Maybe;
 use cairo_lang_semantic as semantic;
 use itertools::{chain, zip_eq, Itertools};
+use semantic::items::functions::{
+    ConcreteFunctionWithBody, GenericFunctionId, GenericFunctionWithBodyId,
+};
 use semantic::{ConcreteFunctionWithBodyId, TypeId};
 
-use crate::db::LoweringGroup;
+use crate::db::{ConcreteSCCRepresentative, LoweringGroup};
+use crate::graph_algorithms::strongly_connected_components::concrete_function_with_body_scc;
 use crate::lower::context::{LoweringContext, LoweringContextBuilder, VarRequest};
 use crate::{BlockId, FlatBlockEnd, FlatLowered, MatchInfo, Statement, VariableId};
 
@@ -177,4 +181,149 @@ fn lower_block_implicits(ctx: &mut Context<'_>, block_id: BlockId) -> Maybe<()> 
         lower_block_implicits(ctx, block_id)?;
     }
     Ok(())
+}
+
+// =========== Query implementations ===========
+
+/// Query implementation of [crate::db::LoweringGroup::function_scc_explicit_implicits].
+pub fn function_scc_explicit_implicits(
+    db: &dyn LoweringGroup,
+    function: ConcreteSCCRepresentative,
+) -> Maybe<HashSet<TypeId>> {
+    let scc = concrete_function_with_body_scc(db, function.0);
+    let mut explicit_implicits = HashSet::new();
+    for func in scc {
+        let current_implicits: HashSet<TypeId> = match func.generic_function(db.upcast()) {
+            GenericFunctionWithBodyId::Free(free_function) => {
+                db.free_function_declaration_implicits(free_function)?.into_iter().collect()
+            }
+            GenericFunctionWithBodyId::Impl(concrete_impl_generic_function) => db
+                .impl_function_declaration_implicits(concrete_impl_generic_function.function)?
+                .into_iter()
+                .collect(),
+        };
+        explicit_implicits.extend(current_implicits);
+    }
+    Ok(explicit_implicits)
+}
+
+/// Query implementation of [crate::db::LoweringGroup::function_all_implicits].
+pub fn function_all_implicits(
+    db: &dyn LoweringGroup,
+    function: semantic::FunctionId,
+) -> Maybe<Vec<TypeId>> {
+    let concrete_function = function.get_concrete(db.upcast());
+    match concrete_function.generic_function {
+        GenericFunctionId::Free(free_function) => {
+            let concrete_with_body =
+                db.intern_concrete_function_with_body(ConcreteFunctionWithBody {
+                    generic_function: GenericFunctionWithBodyId::Free(free_function),
+                    generic_args: concrete_function.generic_args,
+                });
+            db.concrete_function_with_body_all_implicits_vec(concrete_with_body)
+        }
+        GenericFunctionId::Impl(impl_generic_function) => {
+            let Some(generic_with_body) =
+                impl_generic_function.to_generic_with_body(db.upcast())?
+                else {
+                    unreachable!();
+                };
+            let concrete_with_body =
+                db.intern_concrete_function_with_body(ConcreteFunctionWithBody {
+                    generic_function: generic_with_body,
+                    generic_args: concrete_function.generic_args,
+                });
+            db.concrete_function_with_body_all_implicits_vec(concrete_with_body)
+        }
+        GenericFunctionId::Extern(extern_function) => {
+            db.extern_function_declaration_implicits(extern_function)
+        }
+        GenericFunctionId::Trait(_) => unreachable!(),
+    }
+}
+
+/// Query implementation of [crate::db::LoweringGroup::concrete_function_with_body_all_implicits].
+pub fn concrete_function_with_body_all_implicits(
+    db: &dyn LoweringGroup,
+    function: ConcreteFunctionWithBodyId,
+) -> Maybe<HashSet<TypeId>> {
+    // Find the SCC representative.
+    let scc_representative = db.concrete_function_with_body_scc_representative(function);
+
+    // Start with the explicit implicits of the SCC.
+    let mut all_implicits = db.function_scc_explicit_implicits(scc_representative.clone())?;
+
+    let direct_callees = db.concrete_function_with_body_direct_callees(function)?;
+    // For each direct callee, add its implicits.
+    for direct_callee in direct_callees {
+        let generic_function = direct_callee.generic_function;
+        let current_implicits = match generic_function {
+            GenericFunctionId::Free(free_function) => {
+                // For a free function, call this method recursively. To avoid cycles, first
+                // check that the callee is not in this function's SCC.
+                let concrete_with_body =
+                    db.intern_concrete_function_with_body(ConcreteFunctionWithBody {
+                        generic_function: GenericFunctionWithBodyId::Free(free_function),
+                        generic_args: direct_callee.generic_args,
+                    });
+                let direct_callee_representative =
+                    db.concrete_function_with_body_scc_representative(concrete_with_body);
+                if direct_callee_representative == scc_representative {
+                    // We already have the implicits of this SCC - do nothing.
+                    continue;
+                }
+                db.concrete_function_with_body_all_implicits(direct_callee_representative.0)?
+            }
+            GenericFunctionId::Impl(impl_generic_function) => {
+                let Some(generic_with_body) =
+                    impl_generic_function.to_generic_with_body(db.upcast())?
+                    else {
+                        unreachable!();
+                    };
+                // For an impl function, call this method recursively. To avoid cycles, first
+                // check that the callee is not in this function's SCC.
+                let concrete_with_body =
+                    db.intern_concrete_function_with_body(ConcreteFunctionWithBody {
+                        generic_function: generic_with_body,
+                        generic_args: direct_callee.generic_args,
+                    });
+                let direct_callee_representative =
+                    db.concrete_function_with_body_scc_representative(concrete_with_body);
+                if direct_callee_representative == scc_representative {
+                    // We already have the implicits of this SCC - do nothing.
+                    continue;
+                }
+                db.concrete_function_with_body_all_implicits(direct_callee_representative.0)?
+            }
+            GenericFunctionId::Extern(extern_function) => {
+                // All implicits of a libfunc are explicit implicits.
+                db.extern_function_declaration_implicits(extern_function)?.into_iter().collect()
+            }
+            GenericFunctionId::Trait(_) => unreachable!(),
+        };
+        all_implicits.extend(&current_implicits);
+    }
+    Ok(all_implicits)
+}
+
+/// Query implementation of
+/// [crate::db::LoweringGroup::concrete_function_with_body_all_implicits_vec].
+pub fn concrete_function_with_body_all_implicits_vec(
+    db: &dyn LoweringGroup,
+    function: ConcreteFunctionWithBodyId,
+) -> Maybe<Vec<TypeId>> {
+    let implicits_set = db.concrete_function_with_body_all_implicits(function)?;
+    let mut implicits_vec = implicits_set.into_iter().collect_vec();
+
+    let semantic_db = db.upcast();
+    let precedence = db.implicit_precedence();
+    implicits_vec.sort_by_cached_key(|type_id| {
+        if let Some(idx) = precedence.iter().position(|item| item == type_id) {
+            return (idx, "".to_string());
+        }
+
+        (precedence.len(), type_id.format(semantic_db))
+    });
+
+    Ok(implicits_vec)
 }
